Playback speed added to animation animators. Used in MettaurSP and Wave SP. PA system updated to print all chips similar to MMBN6. Refactored the PA system to be more accurate and allow for chip matches to be in any position. HP draws on top of scene for all entities. ICE tiles slide Players without FloatShoe enabled (note: still too fast. find way to delay slide). TODO: fix mettaur ui flicker bug. might be related to origin positions. TODO: fix text scrolling bug

+++ BattleNetwork/bnNaviRegistration.cpp
@@ -154,7 +154,7 @@ NaviRegistration & NaviRegistration::GetInstance()
 
 NaviRegistration::~NaviRegistration()
 {
-  for (int i = 0; i < Size(); i++) {
+  for (int i = 0; i < (int)Size(); i++) {
     delete roster[i];
   }
 
@@ -176,7 +176,7 @@ void NaviRegistration::Register(const NaviInfo * info)
 
 const NaviRegistration::NaviInfo & NaviRegistration::At(int index)
 {
-  if (index < 0 || index >= Size())
+  if (index < 0 || index >= (int)Size())
     throw std::runtime_error("Roster index out of bounds");
 
   return *(roster.at(index));
@@ -184,12 +184,12 @@ const NaviRegistration::NaviInfo & NaviRegistration::At(int index)
 
 const unsigned NaviRegistration::Size()
 {
-  return roster.size();
+  return (unsigned)roster.size();
 }
 
 void NaviRegistration::LoadAllNavis(std::atomic<int>& progress)
 {
-  for (int i = 0; i < Size(); i++) {
+  for (int i = 0; i < (int)Size(); i++) {
     roster[i]->loadNaviClass();
 
     Logger::GetMutex()->lock();
